Refactoring, moved classes to separate files
